+++ Netable/Example/Requests/GraphQL/GetAllPostsQuery.swift
@@ -15,7 +15,6 @@ struct GetAllPostsResponse: Decodable {
 
 struct GetAllPostsQuery: GraphQLRequest {
 
-
     typealias Parameters = Empty
     typealias RawResource = GetAllPostsResponse
     typealias FinalResource = [Post]
