+++ Netable/Example/Requests/LoginRequest.swift
@@ -32,4 +32,3 @@ struct LoginRequest: Request {
         ["email"]
     }
 }
-
