--- Netable/Example/Requests/SuperSimpleRequest.swift
@@ -1,23 +1,0 @@
-//
-//  SuperSimpleRequest.swift
-//  Example
-//
-//  Created by Amy Oulton on 2022-12-14.
-//  Copyright © 2022 Steamclock Software. All rights reserved.
-//
-
-import Foundation
-
-
-
-//struct GetVersionRequest: Request {
-//    typealias Parameters = Empty
-//    typealias RawResource = Version
-//
-//    var method: HTTPMethod { return .get }
-//    var jsonKeyDecodingStrategy: JSONDecoder.KeyDecodingStrategy { return .convertFromKebabCase }
-//
-//
-//    var path = "version"
-//
-//}
