+++ Netable/Example/Views/HomeView.swift
@@ -15,36 +15,37 @@ struct HomeView: View {
     var body: some View {
         NavigationView {
             VStack {
-                if viewModel.user != nil {
-                    ScrollView {
-                        HStack {
-                            Text("Cat Diary")
-                                .font(.largeTitle)
-                            Spacer()
-                            NavigationLink(destination: CreatePostView(viewModel: viewModel)) {
-                                Text("New Post")
-                            }.padding()
-                            .background(.blue)
-                            .cornerRadius(10)
-                            .foregroundColor(.white)
-                        }
-                        
-                        if let posts = viewModel.posts {
-                        ForEach(posts, id: \.self) { post in
-                                VStack(alignment: .leading) {
-                                    Text(post.title)
-                                        .font(.title2)
-                                        .padding(.bottom, 4)
-                                    Text(post.content)
-                                    Divider()
-                                }
+                ScrollView {
+                    HStack {
+                        Text("Cat Diary")
+                            .font(.largeTitle)
+                        Spacer()
+                        NavigationLink(destination: CreatePostView(viewModel: viewModel)) {
+                            Text("New Post")
+                        }.padding()
+                        .background(.blue)
+                        .cornerRadius(10)
+                        .foregroundColor(.white)
+                    }
+                    if let posts = viewModel.posts {
+                    ForEach(posts, id: \.self) { post in
+                            VStack(alignment: .leading) {
+                                Text(post.title)
+                                    .font(.title2)
+                                    .padding(.bottom, 4)
+                                Text(post.content)
+                                Divider()
                             }
                         }
                     }
-                } else {
-                    EmptyView()
                 }
             }.padding()
+            .onAppear {
+                viewModel.bindViewModel()
+            }
+            .onDisappear {
+                viewModel.unbindViewModel()
+            }
         }
-    }   
+    }
 }
